Split the models into a package, one file per domain
app/models.py held five classes and would have held about ten after the
catalog and the trip slices, so the move is made now while it is cheap
rather than later as a large rename commit tangled with real changes.
The layout follows the domain slices: stufflist (dropped once the real
API routes land) and household, with catalog and trips to come.

Every module must be reexported by app/models/__init__.py. alembic/env.py
imports app.models and reads Base.metadata right after: a module missing
from __init__.py leaves its tables out of the metadata, and the next
autogenerate silently emits drop_table for them. Nothing else would catch
that, so test_models_registry imports every submodule of the package and
asserts the metadata did not grow — verified to fail against a model
module left unexported.

test_models.py is renamed to test_household_models.py, matching the same
per-domain split.

The ER diagram and an autogenerate run are both unchanged, confirming the
move touched no metadata.
