+++ app/models/__init__.py
@@ -1,0 +1,11 @@
+from app.models.household import Household, HouseholdMember, HouseholdRole, Person, User
+from app.models.stufflist import StuffList
+
+__all__ = [
+    "Household",
+    "HouseholdMember",
+    "HouseholdRole",
+    "Person",
+    "StuffList",
+    "User",
+]
