+++ app/models/household.py
@@ -7,14 +7,6 @@ from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.database import Base
 
 
-class StuffList(Base):
-    __tablename__ = "stufflist"
-    __table_args__ = {"comment": "A list of stuff the user wants to keep track of"}
-
-    id: Mapped[int] = mapped_column(primary_key=True, comment="Surrogate primary key")
-    name: Mapped[str] = mapped_column(index=True, comment="Display name given by the user")
-
-
 class HouseholdRole(enum.StrEnum):
     owner = "owner"
     member = "member"
