+++ tests/test_household_models.py
